Update group id.
