+++ pom.xml
@@ -24,18 +24,19 @@
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>2.4.10</version>
     </parent>
-    <groupId>com.zerg</groupId>
+    <groupId>com.zergclan</groupId>
     <artifactId>wormhole</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>${wormhole.sersion}</version>
     <packaging>pom</packaging>
     <name>${project.artifactId}</name>
 
     <properties>
+        <wormhole.sersion>0.1.0-SNAPSHOT</wormhole.sersion>
         <java.version>1.8</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <!-- tools -->
         <disruptor.version>3.4.4</disruptor.version>
         <mybatis-spring-boot-starter.version>2.1.4</mybatis-spring-boot-starter.version>
-
         <!-- test -->
         <junit5.version>5.7.2</junit5.version>
         <!-- maven-plugin -->
@@ -146,7 +147,8 @@
                             <ignoreMethodAnnotation>lombok.Generated</ignoreMethodAnnotation>
                         </ignoreMethodAnnotations>
                         <excludes>
-                            <exclude>com/zerg/wormhole/etl/*.class</exclude>
+                            <exclude>com/zergclan/wormhole/**/*Test.class</exclude>
+                            <exclude>com/zergclan/wormhole/test/**/*.class</exclude>
                         </excludes>
                     </instrumentation>
                 </configuration>
