--- wormhole-features/wormhole-web/src/main/java/com/zerg/wormhole/web/contorller/DiscoveryController.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.zerg.wormhole.web.contorller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * Service discovery controller.
- */
-@RestController
-public final class DiscoveryController {
-    
-    private static final String SERVER_STATUS_UP = "UP";
-    
-    /**
-     * Get application run status.
-     *
-     * @return run status
-     */
-    @GetMapping("/status")
-    public String status() {
-        return SERVER_STATUS_UP;
-    }
-}
